// vcure-frontend/src/types/onboarding.ts
export type RegionalCuisine =
  | "ANDHRA"
  | "TELANGANA"
  | "TAMIL_NADU"
  | "KARNATAKA"
  | "KERALA"
  | "MAHARASHTRA"
  | "NORTH_INDIAN"
  | "BENGALI"
  | "OTHER";